added platform interface and web implementation
