+++ cached_network_image/lib/src/image_provider/_image_loader.dart
@@ -1,0 +1,76 @@
+import 'dart:async';
+import 'dart:ui' as ui;
+
+import 'package:cached_network_image_platform_interface/cached_network_image_platform_interface.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+
+import 'package:cached_network_image_platform_interface'
+        '/cached_network_image_platform_interface.dart' as platform
+    show ImageLoader;
+import 'package:cached_network_image_platform_interface'
+        '/cached_network_image_platform_interface.dart'
+    show ImageRenderMethodForWeb;
+
+class ImageLoader implements platform.ImageLoader {
+  @override
+  Stream<ui.Codec> loadAsync(
+    String url,
+    String? cacheKey,
+    StreamController<ImageChunkEvent> chunkEvents,
+    DecoderCallback decode,
+    BaseCacheManager cacheManager,
+    int? maxHeight,
+    int? maxWidth,
+    Map<String, String>? headers,
+    Function()? errorListener,
+    ImageRenderMethodForWeb imageRenderMethodForWeb,
+      Function() evictImage,
+  ) async* {
+    try {
+      assert(
+          cacheManager is ImageCacheManager ||
+              (maxWidth == null && maxHeight == null),
+          'To resize the image with a CacheManager the '
+          'CacheManager needs to be an ImageCacheManager. maxWidth and '
+          'maxHeight will be ignored when a normal CacheManager is used.');
+
+      var stream = cacheManager is ImageCacheManager
+          ? cacheManager.getImageFile(url,
+              maxHeight: maxHeight,
+              maxWidth: maxWidth,
+              withProgress: true,
+              headers: headers,
+              key: cacheKey)
+          : cacheManager.getFileStream(url,
+              withProgress: true, headers: headers, key: cacheKey);
+
+      await for (var result in stream) {
+        if (result is DownloadProgress) {
+          chunkEvents.add(ImageChunkEvent(
+            cumulativeBytesLoaded: result.downloaded,
+            expectedTotalBytes: result.totalSize,
+          ));
+        }
+        if (result is FileInfo) {
+          var file = result.file;
+          var bytes = await file.readAsBytes();
+          var decoded = await decode(bytes);
+          yield decoded;
+        }
+      }
+    } catch (e) {
+      // Depending on where the exception was thrown, the image cache may not
+      // have had a chance to track the key in the cache at all.
+      // Schedule a microtask to give the cache a chance to add the key.
+      scheduleMicrotask(() {
+        evictImage();
+      });
+
+      errorListener?.call();
+      rethrow;
+    } finally {
+      await chunkEvents.close();
+    }
+  }
+}
