+++ cached_network_image_web/lib/cached_network_image_web.dart
@@ -1,0 +1,101 @@
+library cached_network_image_web;
+
+import 'dart:async';
+import 'dart:ui' as ui;
+
+import 'package:cached_network_image_platform_interface'
+        '/cached_network_image_platform_interface.dart' as platform
+    show ImageLoader;
+import 'package:cached_network_image_platform_interface'
+        '/cached_network_image_platform_interface.dart'
+    show ImageRenderMethodForWeb;
+import 'package:flutter/material.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+
+class ImageLoader implements platform.ImageLoader {
+  Stream<ui.Codec> loadAsync(
+    String url,
+    String? cacheKey,
+    StreamController<ImageChunkEvent> chunkEvents,
+    DecoderCallback decode,
+    BaseCacheManager cacheManager,
+    int? maxHeight,
+    int? maxWidth,
+    Map<String, String>? headers,
+    Function()? errorListener,
+    ImageRenderMethodForWeb imageRenderMethodForWeb,
+    Function() evictImage,
+  ) {
+    switch (imageRenderMethodForWeb) {
+      case ImageRenderMethodForWeb.HttpGet:
+        return _loadAsyncHttpGet(url, cacheKey, chunkEvents, decode,
+            cacheManager, maxHeight, maxWidth, headers, errorListener,
+          evictImage,);
+      case ImageRenderMethodForWeb.HtmlImage:
+        return loadAsyncHtmlImage(url, chunkEvents, decode).asStream();
+    }
+  }
+
+  Stream<ui.Codec> _loadAsyncHttpGet(
+    String url,
+    String? cacheKey,
+    StreamController<ImageChunkEvent> chunkEvents,
+    DecoderCallback decode,
+    BaseCacheManager cacheManager,
+    int? maxHeight,
+    int? maxWidth,
+    Map<String, String>? headers,
+    Function()? errorListener,
+      Function() evictImage,
+  ) async* {
+    try {
+      await for (var result in cacheManager.getFileStream(url,
+          withProgress: true, headers: headers)) {
+        if (result is DownloadProgress) {
+          chunkEvents.add(ImageChunkEvent(
+            cumulativeBytesLoaded: result.downloaded,
+            expectedTotalBytes: result.totalSize,
+          ));
+        }
+        if (result is FileInfo) {
+          var file = result.file;
+          var bytes = await file.readAsBytes();
+          var decoded = await decode(bytes);
+          yield decoded;
+        }
+      }
+    } catch (e) {
+      // Depending on where the exception was thrown, the image cache may not
+      // have had a chance to track the key in the cache at all.
+      // Schedule a microtask to give the cache a chance to add the key.
+      scheduleMicrotask(() {
+        evictImage();
+      });
+
+      errorListener?.call();
+      rethrow;
+    } finally {
+      await chunkEvents.close();
+    }
+  }
+
+  Future<ui.Codec> loadAsyncHtmlImage(
+    String url,
+    StreamController<ImageChunkEvent> chunkEvents,
+    DecoderCallback decode,
+  ) {
+    final resolved = Uri.base.resolve(url);
+
+    return ui.webOnlyInstantiateImageCodecFromUrl( // ignore: undefined_function
+      resolved,
+      chunkCallback: (int bytes, int total) {
+        chunkEvents.add(
+          ImageChunkEvent(
+            cumulativeBytesLoaded: bytes,
+            expectedTotalBytes: total,
+          ),
+        );
+      },
+    ) as Future<ui.Codec>; // ignore: undefined_function
+  }
+}
